fix: Move serialization related code to folder
