+++ source/VMelnalksnis.NordigenDotNet/Institutions/InstitutionClient.cs
@@ -9,6 +9,8 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
+using VMelnalksnis.NordigenDotNet.Serialization;
+
 namespace VMelnalksnis.NordigenDotNet.Institutions;
 
 /// <inheritdoc />
