+++ source/VMelnalksnis.NordigenDotNet/Requisitions/RequisitionClient.cs
@@ -10,6 +10,8 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
+using VMelnalksnis.NordigenDotNet.Serialization;
+
 namespace VMelnalksnis.NordigenDotNet.Requisitions;
 
 /// <inheritdoc />
