+++ source/VMelnalksnis.NordigenDotNet/Serialization/HttpClientExtensions.cs
@@ -10,7 +10,7 @@ using System.Text.Json.Serialization.Metadata;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace VMelnalksnis.NordigenDotNet;
+namespace VMelnalksnis.NordigenDotNet.Serialization;
 
 internal static class HttpClientExtensions
 {
