+++ source/VMelnalksnis.NordigenDotNet/Serialization/NordigenJsonSerializerOptions.cs
@@ -8,7 +8,7 @@ using System.Text.Json.Serialization;
 using NodaTime;
 using NodaTime.Serialization.SystemTextJson;
 
-namespace VMelnalksnis.NordigenDotNet;
+namespace VMelnalksnis.NordigenDotNet.Serialization;
 
 /// <summary><see cref="JsonSerializerOptions"/> for <see cref="INordigenClient"/>.</summary>
 public sealed class NordigenJsonSerializerOptions
