+++ source/VMelnalksnis.NordigenDotNet/Serialization/YearMonthDayCalendar.cs
